[CONFIG] Temporarily rename meseenger icon
