+++ src/assets/ui/icons/navbar/drop-downs/index.ts
@@ -1,2 +1,2 @@
-export { MessengerIcon } from './MessengerIcon';
+export { MessengerIcon } from './messenger-icon-temp';
 export { NotificationsIcon } from './Notificationsicon';
