+++ src/assets/ui/icons/navbar/drop-downs/messenger-icon-temp.tsx
